+++ flowvision/models/res_mlp.py
@@ -12,16 +12,17 @@ from .registry import ModelCreator
 
 
 model_urls = {
-    "resmlp_12": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_12.zip",
-    "resmlp_12_dist": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_12_dist.zip",
-    "resmlp_24": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_24.zip",
-    "resmlp_24_dist": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_24_dist.zip",
-    "resmlp_24_dino": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_24_dino.zip",
-    "resmlp_36": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_36.zip",
-    "resmlp_36_dist": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_36_dist.zip",
-    "resmlpB_24": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlpB_24.zip",
-    "resmlpB_24_in22k": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlpB_24_in22k.zip",
-    "resmlpB_24_dist": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlpB_24_dist.zip",
+    "resmlp_12_224": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_12.zip",
+    "resmlp_12_distilled_224": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_12_dist.zip",
+    "resmlp_12_224_dino": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_12_224_dino.zip",
+    "resmlp_24_224": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_24.zip",
+    "resmlp_24_distilled_224": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_24_dist.zip",
+    "resmlp_24_224_dino": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_24_224_dino.zip",
+    "resmlp_36_224": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_36.zip",
+    "resmlp_36_distilled_224": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlp_36_dist.zip",
+    "resmlp_big_24_224": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlpB_24.zip",
+    "resmlp_big_24_224_in22k_to_1k": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlpB_24_in22k.zip",
+    "resmlp_big_24_distilled_224": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/ResMLP/resmlpB_24_dist.zip",
 }
 
 
@@ -164,7 +165,7 @@ def _create_resmlp(arch, pretrained=False, progress=True, **model_kwargs):
 
 
 @ModelCreator.register_model
-def resmlp_12(pretrained=False, progress=True, **kwargs):
+def resmlp_12_224(pretrained=False, progress=True, **kwargs):
     """
     Constructs the ResMLP-12 model.
 
@@ -180,7 +181,7 @@ def resmlp_12(pretrained=False, progress=True, **kwargs):
     .. code-block:: python
 
         >>> import flowvision
-        >>> resmlp_12 = flowvision.models.resmlp_12(pretrained=False, progress=True)
+        >>> resmlp_12_224 = flowvision.models.resmlp_12_224(pretrained=False, progress=True)
 
     """
     model_kwargs = dict(
@@ -192,12 +193,12 @@ def resmlp_12(pretrained=False, progress=True, **kwargs):
         **kwargs,
     )
     return _create_resmlp(
-        "resmlp_12", pretrained=pretrained, progress=progress, **model_kwargs
+        "resmlp_12_224", pretrained=pretrained, progress=progress, **model_kwargs
     )
 
 
 @ModelCreator.register_model
-def resmlp_12_dist(pretrained=False, progress=True, **kwargs):
+def resmlp_12_distilled_224(pretrained=False, progress=True, **kwargs):
     """
     Constructs the ResMLP-12 model with distillation.
 
@@ -214,7 +215,7 @@ def resmlp_12_dist(pretrained=False, progress=True, **kwargs):
     .. code-block:: python
 
         >>> import flowvision
-        >>> resmlp_12_dist = flowvision.models.resmlp_12_dist(pretrained=False, progress=True)
+        >>> resmlp_12_distilled_224 = flowvision.models.resmlp_12_distilled_224(pretrained=False, progress=True)
 
     """
     model_kwargs = dict(
@@ -226,12 +227,49 @@ def resmlp_12_dist(pretrained=False, progress=True, **kwargs):
         **kwargs,
     )
     return _create_resmlp(
-        "resmlp_12_dist", pretrained=pretrained, progress=progress, **model_kwargs
+        "resmlp_12_distilled_224",
+        pretrained=pretrained,
+        progress=progress,
+        **model_kwargs,
     )
 
 
 @ModelCreator.register_model
-def resmlp_24(pretrained=False, progress=True, **kwargs):
+def resmlp_12_224_dino(pretrained=False, progress=True, **kwargs):
+    """
+    Constructs the ResMLP-12 model trained under DINO proposed in `"Emerging Properties in Self-Supervised Vision Transformers" <https://arxiv.org/abs/2104.14294>`_.
+
+    .. note::
+        ResMLP-12 model with distillation from `"ResMLP: Feedforward networks for image classification with data-efficient training" <https://arxiv.org/pdf/2105.03404.pdf>`_.
+        Note that this model is the same as resmlp_12 but the pretrained weight is different.
+
+    Args:
+        pretrained (bool): Whether to download the pre-trained model on ImageNet. Default: ``False``
+        progress (bool): If True, displays a progress bar of the download to stderr. Default: ``True``
+
+    For example:
+
+    .. code-block:: python
+
+        >>> import flowvision
+        >>> resmlp_12_224_dino = flowvision.models.resmlp_12_224_dino(pretrained=False, progress=True)
+
+    """
+    model_kwargs = dict(
+        patch_size=16,
+        embed_dim=384,
+        depth=12,
+        Patch_layer=PatchEmbed,
+        init_scale=0.1,
+        **kwargs,
+    )
+    return _create_resmlp(
+        "resmlp_12_224_dino", pretrained=pretrained, progress=progress, **model_kwargs
+    )
+
+
+@ModelCreator.register_model
+def resmlp_24_224(pretrained=False, progress=True, **kwargs):
     """
     Constructs the ResMLP-24 model.
 
@@ -247,7 +285,7 @@ def resmlp_24(pretrained=False, progress=True, **kwargs):
     .. code-block:: python
 
         >>> import flowvision
-        >>> resmlp_24 = flowvision.models.resmlp_24(pretrained=False, progress=True)
+        >>> resmlp_24_224 = flowvision.models.resmlp_24_224(pretrained=False, progress=True)
 
     """
     model_kwargs = dict(
@@ -259,12 +297,12 @@ def resmlp_24(pretrained=False, progress=True, **kwargs):
         **kwargs,
     )
     return _create_resmlp(
-        "resmlp_24", pretrained=pretrained, progress=progress, **model_kwargs
+        "resmlp_24_224", pretrained=pretrained, progress=progress, **model_kwargs
     )
 
 
 @ModelCreator.register_model
-def resmlp_24_dist(pretrained=False, progress=True, **kwargs):
+def resmlp_24_distilled_224(pretrained=False, progress=True, **kwargs):
     """
     Constructs the ResMLP-24 model with distillation.
 
@@ -281,7 +319,7 @@ def resmlp_24_dist(pretrained=False, progress=True, **kwargs):
     .. code-block:: python
 
         >>> import flowvision
-        >>> resmlp_24_dist = flowvision.models.resmlp_24_dist(pretrained=False, progress=True)
+        >>> resmlp_24_distilled_224 = flowvision.models.resmlp_24_distilled_224(pretrained=False, progress=True)
 
     """
     model_kwargs = dict(
@@ -293,12 +331,15 @@ def resmlp_24_dist(pretrained=False, progress=True, **kwargs):
         **kwargs,
     )
     return _create_resmlp(
-        "resmlp_24_dist", pretrained=pretrained, progress=progress, **model_kwargs
+        "resmlp_24_distilled_224",
+        pretrained=pretrained,
+        progress=progress,
+        **model_kwargs,
     )
 
 
 @ModelCreator.register_model
-def resmlp_24_dino(pretrained=False, progress=True, **kwargs):
+def resmlp_24_224_dino(pretrained=False, progress=True, **kwargs):
     """
     Constructs the ResMLP-24 model trained under DINO proposed in `"Emerging Properties in Self-Supervised Vision Transformers" <https://arxiv.org/abs/2104.14294>`_.
 
@@ -315,7 +356,7 @@ def resmlp_24_dino(pretrained=False, progress=True, **kwargs):
     .. code-block:: python
 
         >>> import flowvision
-        >>> resmlp_24_dino = flowvision.models.resmlp_24_dino(pretrained=False, progress=True)
+        >>> resmlp_24_224_dino = flowvision.models.resmlp_24_224_dino(pretrained=False, progress=True)
 
     """
     model_kwargs = dict(
@@ -327,12 +368,12 @@ def resmlp_24_dino(pretrained=False, progress=True, **kwargs):
         **kwargs,
     )
     return _create_resmlp(
-        "resmlp_24_dino", pretrained=pretrained, progress=progress, **model_kwargs
+        "resmlp_24_dino_224", pretrained=pretrained, progress=progress, **model_kwargs
     )
 
 
 @ModelCreator.register_model
-def resmlp_36(pretrained=False, progress=True, **kwargs):
+def resmlp_36_224(pretrained=False, progress=True, **kwargs):
     """
     Constructs the ResMLP-36 model.
 
@@ -348,7 +389,7 @@ def resmlp_36(pretrained=False, progress=True, **kwargs):
     .. code-block:: python
 
         >>> import flowvision
-        >>> resmlp_36 = flowvision.models.resmlp_36(pretrained=False, progress=True)
+        >>> resmlp_36_224 = flowvision.models.resmlp_36_224(pretrained=False, progress=True)
 
     """
     model_kwargs = dict(
@@ -360,12 +401,12 @@ def resmlp_36(pretrained=False, progress=True, **kwargs):
         **kwargs,
     )
     return _create_resmlp(
-        "resmlp_36", pretrained=pretrained, progress=progress, **model_kwargs
+        "resmlp_36_224", pretrained=pretrained, progress=progress, **model_kwargs
     )
 
 
 @ModelCreator.register_model
-def resmlp_36_dist(pretrained=False, progress=True, **kwargs):
+def resmlp_36_distilled_224(pretrained=False, progress=True, **kwargs):
     """
     Constructs the ResMLP-36 model with distillation.
 
@@ -382,7 +423,7 @@ def resmlp_36_dist(pretrained=False, progress=True, **kwargs):
     .. code-block:: python
 
         >>> import flowvision
-        >>> resmlp_36_dist = flowvision.models.resmlp_36_dist(pretrained=False, progress=True)
+        >>> resmlp_36_distilled_224 = flowvision.models.resmlp_36_distilled_224(pretrained=False, progress=True)
 
     """
     model_kwargs = dict(
@@ -394,17 +435,20 @@ def resmlp_36_dist(pretrained=False, progress=True, **kwargs):
         **kwargs,
     )
     return _create_resmlp(
-        "resmlp_36_dist", pretrained=pretrained, progress=progress, **model_kwargs
+        "resmlp_36_distilled_224",
+        pretrained=pretrained,
+        progress=progress,
+        **model_kwargs,
     )
 
 
 @ModelCreator.register_model
-def resmlpB_24(pretrained=False, progress=True, **kwargs):
+def resmlp_big_24_224(pretrained=False, progress=True, **kwargs):
     """
-    Constructs the ResMLP-B-24 model.
+    Constructs the ResMLP-Big-24 model.
 
     .. note::
-        ResMLP-B-24 model from `"ResMLP: Feedforward networks for image classification with data-efficient training" <https://arxiv.org/pdf/2105.03404.pdf>`_.
+        ResMLP-Big-24 model from `"ResMLP: Feedforward networks for image classification with data-efficient training" <https://arxiv.org/pdf/2105.03404.pdf>`_.
 
     Args:
         pretrained (bool): Whether to download the pre-trained model on ImageNet. Default: ``False``
@@ -415,7 +459,7 @@ def resmlpB_24(pretrained=False, progress=True, **kwargs):
     .. code-block:: python
 
         >>> import flowvision
-        >>> resmlpB_24 = flowvision.models.resmlpB_24(pretrained=False, progress=True)
+        >>> resmlp_big_24_224 = flowvision.models.resmlp_big_24_224(pretrained=False, progress=True)
 
     """
     model_kwargs = dict(
@@ -427,12 +471,12 @@ def resmlpB_24(pretrained=False, progress=True, **kwargs):
         **kwargs,
     )
     return _create_resmlp(
-        "resmlpB_24", pretrained=pretrained, progress=progress, **model_kwargs
+        "resmlp_big_24_224", pretrained=pretrained, progress=progress, **model_kwargs
     )
 
 
 @ModelCreator.register_model
-def resmlpB_24_in22k(pretrained=False, progress=True, **kwargs):
+def resmlp_big_24_224_in22k_to_1k(pretrained=False, progress=True, **kwargs):
     """
     Constructs the ImageNet22k pretrained ResMLP-B-24 model.
 
@@ -449,7 +493,7 @@ def resmlpB_24_in22k(pretrained=False, progress=True, **kwargs):
     .. code-block:: python
 
         >>> import flowvision
-        >>> resmlpB_24_in22k = flowvision.models.resmlpB_24_in22k(pretrained=False, progress=True)
+        >>> resmlp_big_24_224_in22k_to_1k = flowvision.models.resmlp_big_24_224_in22k_to_1k(pretrained=False, progress=True)
 
     """
     model_kwargs = dict(
@@ -461,12 +505,15 @@ def resmlpB_24_in22k(pretrained=False, progress=True, **kwargs):
         **kwargs,
     )
     return _create_resmlp(
-        "resmlpB_24_in22k", pretrained=pretrained, progress=progress, **model_kwargs
+        "resmlp_big_24_224_in22k_to_1k",
+        pretrained=pretrained,
+        progress=progress,
+        **model_kwargs,
     )
 
 
 @ModelCreator.register_model
-def resmlpB_24_dist(pretrained=False, progress=True, **kwargs):
+def resmlp_big_24_distilled_224(pretrained=False, progress=True, **kwargs):
     """
     Constructs the ResMLP-B-24 model with distillation.
 
@@ -483,7 +530,7 @@ def resmlpB_24_dist(pretrained=False, progress=True, **kwargs):
     .. code-block:: python
 
         >>> import flowvision
-        >>> resmlpB_24_dist = flowvision.models.resmlpB_24_dist(pretrained=False, progress=True)
+        >>> resmlp_big_24_distilled_224 = flowvision.models.resmlp_big_24_distilled_224(pretrained=False, progress=True)
 
     """
     model_kwargs = dict(
@@ -495,5 +542,8 @@ def resmlpB_24_dist(pretrained=False, progress=True, **kwargs):
         **kwargs,
     )
     return _create_resmlp(
-        "resmlpB_24_dist", pretrained=pretrained, progress=progress, **model_kwargs
+        "resmlp_big_24_distilled_224",
+        pretrained=pretrained,
+        progress=progress,
+        **model_kwargs,
     )